change the select query to include also the person number

+++ src/main/java/sda/tema/SDA_Tema_4/business/TripService.java
@@ -35,7 +35,7 @@ public class TripService {
                 request.getToDate(),
                 request.getHotelName(),
                 request.getCityName(),
-                request.getNrOfPersons());
+                request.getNrOfPersons().longValue());
 
         if (null != listOfTrips) {
             int listSize = listOfTrips.size();
